* Ignore warnings at end of CPAN.pm output
I moved around the stuff to get the last line from the CPAN.pm
output. On Windows, there was an extra line at the end warning
about YAML not being installed. The final status message was
right above that. Now I look at the output and work backward
until I find a line that isn't a warning.

I added a '-Mblib' to the command that runs the cpan(1) command.
It needs to find App/Cpan.pm and on Windows that wasn't
propogated for some reason.

There are some whitespace oddities here because Komodo went a
little crazy with indention and tabs.